adding xhr-base
