+++ src/document-fetcher.js
@@ -14,23 +14,39 @@
  * limitations under the License.
  */
 
-import {FetchInitDef, fetchPolyfill} from './fetch-polyfill.js';
-import {Services} from './services';
-import {setupInit} from './service/xhr-impl';
+import {FetchInitDef, FetchResponse, fetchPolyfill} from './fetch-polyfill.js';
+import {XhrBase, assertSuccess, setupInit} from './service/xhr-base';
+import {dev} from './log';
+import {isFormDataWrapper} from './form-data-wrapper';
 
-export class DocumentFetcher {
+export class DocumentFetcher extends XhrBase {
   /**
    * Creates an instance of DocumentFetcher.
    * @param {!Window} win
    */
   constructor(win) {
-    /** @type {!Window} */
-    this.win_ = win;
+    super(win);
+  }
 
-    /** @const @private {!./service/xhr-impl.Xhr} */
-    this.xhr_ = Services.xhrFor(this.win_);
+  /**
+   * @override
+   */
+  fetch_(input, init) {
+    dev().assert(typeof input == 'string', 'Only URL supported: %s', input);
 
-    this.xhr_.setFetchOverride(fetchPolyfill);
+    return this.maybeIntercept_(input, init)
+        .then(interceptorResponse => {
+          if (interceptorResponse) {
+            return interceptorResponse;
+          }
+          // After this point, both the native `fetch` and the `fetch` polyfill
+          // will expect a native `FormData` object in the `body` property, so
+          // the native `FormData` object needs to be unwrapped.
+          if (isFormDataWrapper(init.body)) {
+            init.body = init.body.getFormData();
+          }
+          return (fetchPolyfill).apply(null, arguments);
+        });
   }
 
   /**
@@ -38,12 +54,25 @@ export class DocumentFetcher {
    *
    * @param {string} input
    * @param {?FetchInitDef=} opt_init
-   * @return {Response}
+   * @return {!Promise<!Document>}
+   * @ignore
    */
   fetchDocument(input, opt_init) {
     const init = setupInit(opt_init, 'text/html');
     init.responseType = 'document';
-    return this.xhr_.fetch(input, init)
-        .then(response => response.document_());
+    return this.fetchAmpCors_(input, init).then(res => {
+      const response = /**@type {!Response} */ (res);
+      return assertSuccess(response);
+    }).then(res => {
+      const response = /** @type {!FetchResponse} */ (res);
+      dev().assert(response.document_,
+          'Document method not found, make sure you passed fetch polyfill before fetching a document');
+      return response.document_();
+    });
   }
+}
+
+/** @package @VisibleForTesting */
+export function fetchResponseForTesting() {
+  return FetchResponse;
 }
